+++ src/org/moosetechnology/verveineC/plugin/VerveineCParser.java
@@ -30,8 +30,8 @@ import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.Path;
 import org.moosetechnology.verveineC.utils.Constants;
-import org.moosetechnology.verveineC.utils.FileUtil;
 import org.moosetechnology.verveineC.utils.Trace;
+import org.moosetechnology.verveineC.utils.fileAndStream.FileUtil;
 import org.moosetechnology.verveineC.visitors.AbstractIssueReporterVisitor;
 import org.moosetechnology.verveineC.visitors.ErrorVisitor;
 import org.moosetechnology.verveineC.visitors.IncludeVisitor;
