+++ src/org/moosetechnology/verveineC/utils/resolution/StubBinding.java
@@ -1,4 +1,4 @@
-package org.moosetechnology.verveineC.utils;
+package org.moosetechnology.verveineC.utils.resolution;
 
 import java.util.HashMap;
 import java.util.Map;
