+++ src/org/moosetechnology/verveineC/visitors/SignatureBuilderVisitor.java
@@ -11,8 +11,8 @@ import org.eclipse.cdt.core.dom.ast.IASTSimpleDeclaration;
 import org.eclipse.cdt.core.dom.ast.IASTStandardFunctionDeclarator;
 import org.eclipse.cdt.core.dom.ast.gnu.c.ICASTKnRFunctionDeclarator;
 import org.moosetechnology.verveineC.plugin.CDictionary;
-import org.moosetechnology.verveineC.utils.QualifiedName;
 import org.moosetechnology.verveineC.utils.Trace;
+import org.moosetechnology.verveineC.utils.resolution.QualifiedName;
 
 /**
  * A visitor specialized in reconstructing the signature of a method/function.
