+++ src/org/moosetechnology/verveineC/visitors/def/CommentDefVisitor.java
@@ -7,7 +7,7 @@ import org.eclipse.cdt.core.index.IIndex;
 import org.eclipse.cdt.core.model.ITranslationUnit;
 import org.eclipse.core.runtime.CoreException;
 import org.moosetechnology.verveineC.plugin.CDictionary;
-import org.moosetechnology.verveineC.utils.FileUtil;
+import org.moosetechnology.verveineC.utils.fileAndStream.FileUtil;
 import org.moosetechnology.verveineC.visitors.AbstractVisitor;
 
 import eu.synectique.verveine.core.gen.famix.Comment;
