+++ src/org/moosetechnology/verveineC/visitors/ref/AbstractRefVisitor.java
@@ -25,9 +25,9 @@ import org.eclipse.cdt.core.dom.ast.cpp.ICPPTemplateInstance;
 import org.eclipse.cdt.core.dom.ast.gnu.c.ICASTKnRFunctionDeclarator;
 import org.eclipse.cdt.core.index.IIndex;
 import org.moosetechnology.verveineC.plugin.CDictionary;
-import org.moosetechnology.verveineC.utils.QualifiedName;
-import org.moosetechnology.verveineC.utils.StubBinding;
 import org.moosetechnology.verveineC.utils.Trace;
+import org.moosetechnology.verveineC.utils.resolution.QualifiedName;
+import org.moosetechnology.verveineC.utils.resolution.StubBinding;
 import org.moosetechnology.verveineC.visitors.AbstractVisitor;
 
 import eu.synectique.verveine.core.EntityStack;
