+++ tests/verveine/extractor/Cpp/IncludeFilterTest.java
@@ -5,8 +5,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.StringBufferInputStream;
 
-import org.moosetechnology.verveineC.utils.IncludeToLowerFilterStream;
-import org.moosetechnology.verveineC.utils.IncludeWithHExtensionFilterStream;
+import org.moosetechnology.verveineC.utils.fileAndStream.IncludeToLowerFilterStream;
+import org.moosetechnology.verveineC.utils.fileAndStream.IncludeWithHExtensionFilterStream;
 
 @SuppressWarnings("deprecation")
 public class IncludeFilterTest {
